clean up upgrade files
